Allow enumerations to be of any base numeric type (#4)
This allows enumerations to be of any base numeric type.

The downside is that there is no general value that can be used for
"invalid enumeration", so this has to change the interface of EnumOf to
be (T, bool) instead of returning an invalid value. This is arguably
more Go-style anyhow. Plus it is also just generally true that you can't
pick an enumeration value to be a generic "you can't use this" because
you don't know what users may want.

Not necessarily suggesting you take this, this was just easier than
trying to describe what I meant in /r/golang.

The "mtmap.go" is stripped of docs since it is not exported, but it's an
adaptation of [mtmap](https://pkg.go.dev/github.com/thejerf/mtmap). As
per the README itself, it's not something you use as a library exactly,
but best copied & pasted & modified as needed. In this case, the key
type changed quite substantially for your use case. This allows for an
`enums` map that stores everything directly, and strongly-typed.
However, the API described in the docs is still effectively the same.

Some more propagation of "there isn't really an invalid value for enum"
may be useful.

---------

Co-authored-by: Jeremy Bowers <jeremy.bowers@barracuda.com>
Co-authored-by: huykingsofm <44906193+huykingsofm@users.noreply.github.com>